Add an else statement for previous button
Add an else statement for the previous button after the if statement on
line 36 so that it is not shown immediately the page loads.

This means that you can't go back yet it is the first image being shown
hence only show the next button telling the user to go to the next
image.